Move SyncScheduler to its own file.
